Split Enemy into BehaviorEnemy and AIStateEnemy, added colliderboxes for Spiderling and Rachnok
