+++ Server.Reawakened/Entities/Components/BaseSpawnerControllerComp.cs
@@ -5,8 +5,8 @@ using Server.Base.Logging;
 using Server.Base.Timers.Extensions;
 using Server.Base.Timers.Services;
 using Server.Reawakened.Configs;
-using Server.Reawakened.Entities.Entity;
 using Server.Reawakened.Entities.Entity.Enemies;
+using Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
 using Server.Reawakened.Entities.Entity.Utils;
 using Server.Reawakened.Players;
 using Server.Reawakened.Players.Helpers;
