+++ Server.Reawakened/Entities/Entity/Enemies/BehaviorEnemies/EnemyFish.cs
@@ -6,8 +6,8 @@ using Server.Reawakened.Rooms;
 using Server.Reawakened.Rooms.Extensions;
 using Server.Reawakened.Rooms.Models.Entities;
 
-namespace Server.Reawakened.Entities.Entity.Enemies;
-public class EnemyBathog(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
+namespace Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
+public class EnemyFish(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
 {
 
     private float _behaviorEndTime;
@@ -27,13 +27,10 @@ public class EnemyBathog(Room room, string entityId, string prefabName, EnemyCon
         EnemyGlobalProps.Global_BackDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeUpY"));
         EnemyGlobalProps.Global_BackDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeDownY"));
         EnemyGlobalProps.Aggro_AttackBeyondPatrolLine = Convert.ToSingle(BehaviorList.GetGlobalProperty("AttackBeyondPatrolLine"));
-        EnemyGlobalProps.Global_ShootOffsetX = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetX"));
-        EnemyGlobalProps.Global_ShootOffsetY = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetY"));
-        EnemyGlobalProps.Global_ShootingProjectilePrefabName = BehaviorList.GetGlobalProperty("ProjectilePrefabName").ToString();
 
         // Address magic numbers when we get to adding enemy effect mods
         Room.SendSyncEvent(AIInit(1, 1, 1));
-        Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, -1, false));
+        Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, 1, false));
 
         // Set these calls to the xml later. Instead of using hardcoded "Patrol", "Aggro", etc.
         // the XML can just specify which behaviors to use when attacked, when moving, etc.
@@ -63,7 +60,7 @@ public class EnemyBathog(Room room, string entityId, string prefabName, EnemyCon
     {
         base.HandlePatrol();
 
-        DetectPlayers(_offensiveBehavior);
+        DetectPlayers("Aggro");
     }
 
     public override void HandleAggro()
@@ -84,7 +81,7 @@ public class EnemyBathog(Room room, string entityId, string prefabName, EnemyCon
     {
         base.HandleLookAround();
 
-        DetectPlayers(_offensiveBehavior);
+        DetectPlayers("Aggro");
 
         if (Room.Time >= _behaviorEndTime)
         {
@@ -96,20 +93,6 @@ public class EnemyBathog(Room room, string entityId, string prefabName, EnemyCon
 
             AiBehavior = ChangeBehavior("ComeBack");
             AiBehavior.MustDoComeback(AiData);
-        }
-    }
-
-    public override void HandleShooting()
-    {
-        base.HandleShooting();
-
-        if (!AiBehavior.Update(ref AiData, Room.Time))
-        {
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, AiData.Sync_TargetPosX, AiData.Sync_TargetPosY,
-            AiData.Intern_Dir, false));
-
-            AiBehavior = ChangeBehavior("LookAround");
-            _behaviorEndTime = ResetBehaviorTime(Convert.ToSingle(BehaviorList.GetBehaviorStat("LookAround", "lookTime")));
         }
     }
 
@@ -127,7 +110,6 @@ public class EnemyBathog(Room room, string entityId, string prefabName, EnemyCon
     public override void DetectPlayers(string behaviorToRun)
     {
         foreach (var player in Room.Players)
-        {
             if (PlayerInRange(player.Value.TempData.Position, EnemyGlobalProps.Global_DetectionLimitedByPatrolLine))
             {
                 Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(behaviorToRun), string.Empty, player.Value.TempData.Position.X,
@@ -141,6 +123,5 @@ public class EnemyBathog(Room room, string entityId, string prefabName, EnemyCon
 
                 _behaviorEndTime = ResetBehaviorTime(MinBehaviorTime);
             }
-        }
     }
 }
