+++ Server.Reawakened/Entities/Entity/Enemies/BehaviorEnemies/EnemyGeneric.cs
@@ -2,7 +2,7 @@
 using Server.Reawakened.Rooms;
 using Server.Reawakened.Rooms.Models.Entities;
 
-namespace Server.Reawakened.Entities.Entity.Enemies;
+namespace Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
 public class EnemyGeneric(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
 {
 }
