+++ Server.Reawakened/Entities/Entity/Enemies/BehaviorEnemies/EnemyGrenadier.cs
@@ -5,8 +5,8 @@ using Server.Reawakened.Rooms;
 using Server.Reawakened.Rooms.Extensions;
 using Server.Reawakened.Rooms.Models.Entities;
 
-namespace Server.Reawakened.Entities.Entity.Enemies;
-public class EnemyBird(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
+namespace Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
+public class EnemyGrenadier(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
 {
 
     private float _behaviorEndTime;
@@ -15,6 +15,12 @@ public class EnemyBird(Room room, string entityId, string prefabName, EnemyContr
     public override void Initialize()
     {
         base.Initialize();
+        AiData.services = new AIServices
+        {
+            _shoot = new IShoot(),
+            _bomber = new IBomber(),
+            _scan = new IScan()
+        };
 
         MinBehaviorTime = Convert.ToSingle(BehaviorList.GetGlobalProperty("MinBehaviorTime"));
         _offensiveBehavior = Convert.ToString(BehaviorList.GetGlobalProperty("OffensiveBehavior"));
@@ -23,13 +29,12 @@ public class EnemyBird(Room room, string entityId, string prefabName, EnemyContr
         EnemyGlobalProps.Global_FrontDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("FrontDetectionRangeUpY"));
         EnemyGlobalProps.Global_FrontDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("FrontDetectionRangeDownY"));
         EnemyGlobalProps.Global_BackDetectionRangeX = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeX"));
-        EnemyGlobalProps.Global_ShootOffsetX = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetX"));
-        EnemyGlobalProps.Global_ShootOffsetY = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetY"));
-        EnemyGlobalProps.Global_ShootingProjectilePrefabName = BehaviorList.GetGlobalProperty("ProjectilePrefabName").ToString();
+        EnemyGlobalProps.Global_BackDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeUpY"));
+        EnemyGlobalProps.Global_BackDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeDownY"));
 
         // Address magic numbers when we get to adding enemy effect mods
         Room.SendSyncEvent(AIInit(1, 1, 1));
-        Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, -1, false));
+        Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, 1, false));
 
         // Set these calls to the xml later. Instead of using hardcoded "Patrol", "Aggro", etc.
         // the XML can just specify which behaviors to use when attacked, when moving, etc.
@@ -59,17 +64,16 @@ public class EnemyBird(Room room, string entityId, string prefabName, EnemyContr
     {
         base.HandlePatrol();
 
-        DetectPlayers("Shooting");
+        DetectPlayers("Grenadier");
     }
 
-    public override void HandleShooting()
+    public override void HandleGrenadier()
     {
-        base.HandleShooting();
+        base.HandleGrenadier();
 
-        if (!AiBehavior.Update(ref AiData, Room.Time))
+        if (Room.Time >= _behaviorEndTime)
         {
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, AiData.Sync_TargetPosX, AiData.Sync_TargetPosY,
-            AiData.Intern_Dir, false));
+            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, Position.x, Position.y, AiData.Intern_Dir, false));
 
             AiBehavior = ChangeBehavior("LookAround");
             _behaviorEndTime = ResetBehaviorTime(Convert.ToSingle(BehaviorList.GetBehaviorStat("LookAround", "lookTime")));
@@ -80,7 +84,7 @@ public class EnemyBird(Room room, string entityId, string prefabName, EnemyContr
     {
         base.HandleLookAround();
 
-        DetectPlayers("Shooting");
+        DetectPlayers("Grenadier");
 
         if (Room.Time >= _behaviorEndTime)
         {
@@ -92,20 +96,21 @@ public class EnemyBird(Room room, string entityId, string prefabName, EnemyContr
     public override void DetectPlayers(string behaviorToRun)
     {
         foreach (var player in Room.Players)
-        {
             if (PlayerInRange(player.Value.TempData.Position, EnemyGlobalProps.Global_DetectionLimitedByPatrolLine))
             {
                 Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(behaviorToRun), string.Empty, player.Value.TempData.Position.X,
-                    player.Value.TempData.Position.Y, Generic.Patrol_ForceDirectionX, false));
+                    Position.y, Generic.Patrol_ForceDirectionX, false));
 
                 // For some reason, the SyncEvent doesn't initialize these properly, so I just do them here
                 AiData.Sync_TargetPosX = player.Value.TempData.Position.X;
-                AiData.Sync_TargetPosY = player.Value.TempData.Position.Y;
+                AiData.Sync_TargetPosY = Position.y;
 
                 AiBehavior = ChangeBehavior(behaviorToRun);
 
-                _behaviorEndTime = ResetBehaviorTime(MinBehaviorTime);
+                _behaviorEndTime = ResetBehaviorTime(
+                Convert.ToSingle(BehaviorList.GetBehaviorStat("Grenadier", "inTime")) +
+                Convert.ToSingle(BehaviorList.GetBehaviorStat("Grenadier", "loopTime")) +
+                Convert.ToSingle(BehaviorList.GetBehaviorStat("Grenadier", "outTime")));
             }
-        }
     }
 }
