+++ Server.Reawakened/Entities/Entity/Enemies/BehaviorEnemies/EnemyPincer.cs
@@ -5,8 +5,8 @@ using Server.Reawakened.Rooms;
 using Server.Reawakened.Rooms.Extensions;
 using Server.Reawakened.Rooms.Models.Entities;
 
-namespace Server.Reawakened.Entities.Entity.Enemies;
-public class EnemyOrchid(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
+namespace Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
+public class EnemyPincer(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
 {
 
     private float _behaviorEndTime;
@@ -23,9 +23,8 @@ public class EnemyOrchid(Room room, string entityId, string prefabName, EnemyCon
         EnemyGlobalProps.Global_FrontDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("FrontDetectionRangeUpY"));
         EnemyGlobalProps.Global_FrontDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("FrontDetectionRangeDownY"));
         EnemyGlobalProps.Global_BackDetectionRangeX = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeX"));
-        EnemyGlobalProps.Global_ShootOffsetX = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetX"));
-        EnemyGlobalProps.Global_ShootOffsetY = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetY"));
-        EnemyGlobalProps.Global_ShootingProjectilePrefabName = BehaviorList.GetGlobalProperty("ProjectilePrefabName").ToString();
+        EnemyGlobalProps.Global_BackDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeUpY"));
+        EnemyGlobalProps.Global_BackDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeDownY"));
 
         // Address magic numbers when we get to adding enemy effect mods
         Room.SendSyncEvent(AIInit(1, 1, 1));
@@ -52,9 +51,6 @@ public class EnemyOrchid(Room room, string entityId, string prefabName, EnemyCon
             AiBehavior = ChangeBehavior(_offensiveBehavior);
 
             _behaviorEndTime = ResetBehaviorTime(MinBehaviorTime);
-
-            if (Id.Equals("7651"))
-                Console.WriteLine("You pissed me off.");
         }
     }
 
@@ -62,19 +58,20 @@ public class EnemyOrchid(Room room, string entityId, string prefabName, EnemyCon
     {
         base.HandlePatrol();
 
-        DetectPlayers("Shooting");
+        DetectPlayers("Aggro");
     }
 
-    public override void HandleShooting()
+    public override void HandleAggro()
     {
-        base.HandleShooting();
+        base.HandleAggro();
 
         if (!AiBehavior.Update(ref AiData, Room.Time))
         {
+
+            AiBehavior = ChangeBehavior("LookAround");
             Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, AiData.Sync_TargetPosX, AiData.Sync_TargetPosY,
             AiData.Intern_Dir, false));
 
-            AiBehavior = ChangeBehavior("LookAround");
             _behaviorEndTime = ResetBehaviorTime(Convert.ToSingle(BehaviorList.GetBehaviorStat("LookAround", "lookTime")));
         }
     }
@@ -83,33 +80,32 @@ public class EnemyOrchid(Room room, string entityId, string prefabName, EnemyCon
     {
         base.HandleLookAround();
 
-        DetectPlayers("Shooting");
+        DetectPlayers("Aggro");
 
         if (Room.Time >= _behaviorEndTime)
         {
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, AiData.Intern_Dir, false));
-
             AiBehavior = ChangeBehavior("Patrol");
+            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, Generic.Patrol_ForceDirectionX, false));
         }
     }
 
     public override void DetectPlayers(string behaviorToRun)
     {
         foreach (var player in Room.Players)
-        {
             if (PlayerInRange(player.Value.TempData.Position, EnemyGlobalProps.Global_DetectionLimitedByPatrolLine))
             {
+                AiBehavior.Stop(ref AiData);
+
                 Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(behaviorToRun), string.Empty, player.Value.TempData.Position.X,
-                    player.Value.TempData.Position.Y, Generic.Patrol_ForceDirectionX, false));
+                    Position.y, AiData.Intern_Dir, false));
 
                 // For some reason, the SyncEvent doesn't initialize these properly, so I just do them here
                 AiData.Sync_TargetPosX = player.Value.TempData.Position.X;
-                AiData.Sync_TargetPosY = player.Value.TempData.Position.Y;
+                AiData.Sync_TargetPosY = Position.y;
 
                 AiBehavior = ChangeBehavior(behaviorToRun);
 
                 _behaviorEndTime = ResetBehaviorTime(MinBehaviorTime);
             }
-        }
     }
 }
