+++ Server.Reawakened/Entities/Entity/Enemies/BehaviorEnemies/EnemySpider.cs
@@ -5,8 +5,8 @@ using Server.Reawakened.Rooms;
 using Server.Reawakened.Rooms.Extensions;
 using Server.Reawakened.Rooms.Models.Entities;
 
-namespace Server.Reawakened.Entities.Entity.Enemies;
-public class EnemyGrenadier(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
+namespace Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
+public class EnemySpider(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
 {
 
     private float _behaviorEndTime;
@@ -15,12 +15,6 @@ public class EnemyGrenadier(Room room, string entityId, string prefabName, Enemy
     public override void Initialize()
     {
         base.Initialize();
-        AiData.services = new AIServices
-        {
-            _shoot = new IShoot(),
-            _bomber = new IBomber(),
-            _scan = new IScan()
-        };
 
         MinBehaviorTime = Convert.ToSingle(BehaviorList.GetGlobalProperty("MinBehaviorTime"));
         _offensiveBehavior = Convert.ToString(BehaviorList.GetGlobalProperty("OffensiveBehavior"));
@@ -31,6 +25,9 @@ public class EnemyGrenadier(Room room, string entityId, string prefabName, Enemy
         EnemyGlobalProps.Global_BackDetectionRangeX = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeX"));
         EnemyGlobalProps.Global_BackDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeUpY"));
         EnemyGlobalProps.Global_BackDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeDownY"));
+        EnemyGlobalProps.Global_ShootOffsetX = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetX"));
+        EnemyGlobalProps.Global_ShootOffsetY = Convert.ToSingle(BehaviorList.GetGlobalProperty("ShootOffsetY"));
+        EnemyGlobalProps.Global_ShootingProjectilePrefabName = BehaviorList.GetGlobalProperty("ProjectilePrefabName").ToString();
 
         // Address magic numbers when we get to adding enemy effect mods
         Room.SendSyncEvent(AIInit(1, 1, 1));
@@ -64,16 +61,17 @@ public class EnemyGrenadier(Room room, string entityId, string prefabName, Enemy
     {
         base.HandlePatrol();
 
-        DetectPlayers("Grenadier");
+        DetectPlayers(_offensiveBehavior);
     }
 
-    public override void HandleGrenadier()
+    public override void HandleAggro()
     {
-        base.HandleGrenadier();
+        base.HandleAggro();
 
-        if (Room.Time >= _behaviorEndTime)
+        if (!AiBehavior.Update(ref AiData, Room.Time))
         {
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, Position.x, Position.y, AiData.Intern_Dir, false));
+            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, AiData.Sync_TargetPosX, AiData.Sync_TargetPosY,
+            AiData.Intern_Dir, false));
 
             AiBehavior = ChangeBehavior("LookAround");
             _behaviorEndTime = ResetBehaviorTime(Convert.ToSingle(BehaviorList.GetBehaviorStat("LookAround", "lookTime")));
@@ -84,7 +82,7 @@ public class EnemyGrenadier(Room room, string entityId, string prefabName, Enemy
     {
         base.HandleLookAround();
 
-        DetectPlayers("Grenadier");
+        DetectPlayers(_offensiveBehavior);
 
         if (Room.Time >= _behaviorEndTime)
         {
@@ -93,10 +91,23 @@ public class EnemyGrenadier(Room room, string entityId, string prefabName, Enemy
         }
     }
 
+    public override void HandleShooting()
+    {
+        base.HandleShooting();
+
+        if (!AiBehavior.Update(ref AiData, Room.Time))
+        {
+            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("LookAround"), string.Empty, AiData.Sync_TargetPosX, AiData.Sync_TargetPosY,
+            AiData.Intern_Dir, false));
+
+            AiBehavior = ChangeBehavior("LookAround");
+            _behaviorEndTime = ResetBehaviorTime(Convert.ToSingle(BehaviorList.GetBehaviorStat("LookAround", "lookTime")));
+        }
+    }
+
     public override void DetectPlayers(string behaviorToRun)
     {
         foreach (var player in Room.Players)
-        {
             if (PlayerInRange(player.Value.TempData.Position, EnemyGlobalProps.Global_DetectionLimitedByPatrolLine))
             {
                 Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(behaviorToRun), string.Empty, player.Value.TempData.Position.X,
@@ -108,11 +119,7 @@ public class EnemyGrenadier(Room room, string entityId, string prefabName, Enemy
 
                 AiBehavior = ChangeBehavior(behaviorToRun);
 
-                _behaviorEndTime = ResetBehaviorTime(
-                Convert.ToSingle(BehaviorList.GetBehaviorStat("Grenadier", "inTime")) +
-                Convert.ToSingle(BehaviorList.GetBehaviorStat("Grenadier", "loopTime")) +
-                Convert.ToSingle(BehaviorList.GetBehaviorStat("Grenadier", "outTime")));
+                _behaviorEndTime = ResetBehaviorTime(MinBehaviorTime);
             }
-        }
     }
 }
