+++ Server.Reawakened/Entities/Entity/Enemies/BehaviorEnemies/EnemyVespid.cs
@@ -1,12 +1,13 @@
 ﻿using Server.Reawakened.Entities.AIBehavior;
 using Server.Reawakened.Entities.Components;
+using Server.Reawakened.Entities.Entity.Utils;
 using Server.Reawakened.Players;
 using Server.Reawakened.Rooms;
 using Server.Reawakened.Rooms.Extensions;
 using Server.Reawakened.Rooms.Models.Entities;
 
-namespace Server.Reawakened.Entities.Entity.Enemies;
-public class EnemyStomper(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
+namespace Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
+public class EnemyVespid(Room room, string entityId, string prefabName, EnemyControllerComp enemyController, IServiceProvider services) : BehaviorEnemy(room, entityId, prefabName, enemyController, services)
 {
 
     private float _behaviorEndTime;
@@ -26,9 +27,12 @@ public class EnemyStomper(Room room, string entityId, string prefabName, EnemyCo
         EnemyGlobalProps.Global_BackDetectionRangeUpY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeUpY"));
         EnemyGlobalProps.Global_BackDetectionRangeDownY = Convert.ToSingle(BehaviorList.GetGlobalProperty("BackDetectionRangeDownY"));
 
+        AiData.Intern_Dir = Generic.Patrol_ForceDirectionX;
+
         // Address magic numbers when we get to adding enemy effect mods
+        Position.z = 10;
         Room.SendSyncEvent(AIInit(1, 1, 1));
-        Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, 1, false));
+        Room.SendSyncEvent(AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, AiData.Intern_Dir, false));
 
         // Set these calls to the xml later. Instead of using hardcoded "Patrol", "Aggro", etc.
         // the XML can just specify which behaviors to use when attacked, when moving, etc.
@@ -41,8 +45,12 @@ public class EnemyStomper(Room room, string entityId, string prefabName, EnemyCo
 
         if (AiBehavior is not AIBehaviorShooting)
         {
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(_offensiveBehavior), string.Empty, player.TempData.Position.X,
-                    player.TempData.Position.Y, Generic.Patrol_ForceDirectionX, false));
+            var aiEvent = AISyncEventHelper.AIDo(
+                Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(_offensiveBehavior), string.Empty,
+                player.TempData.Position.X, player.TempData.Position.Y, Generic.Patrol_ForceDirectionX, false
+            );
+
+            Room.SendSyncEvent(aiEvent);
 
             // For some reason, the SyncEvent doesn't initialize these properly, so I just do them here
             AiData.Sync_TargetPosX = player.TempData.Position.X;
@@ -58,43 +66,28 @@ public class EnemyStomper(Room room, string entityId, string prefabName, EnemyCo
     {
         base.HandlePatrol();
 
-        DetectPlayers("Aggro");
+        if (Room.Time >= _behaviorEndTime)
+            DetectPlayers("Stinger");
     }
 
-    public override void HandleAggro()
+    public override void HandleStinger()
     {
-        base.HandleAggro();
+        base.HandleStinger();
 
         if (!AiBehavior.Update(ref AiData, Room.Time))
         {
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Stomper"), string.Empty, AiData.Sync_TargetPosX, AiData.Sync_TargetPosY,
-            AiData.Intern_Dir, false));
+            Room.SendSyncEvent(AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, AiData.Intern_Dir, false));
 
-            AiBehavior = ChangeBehavior("Stomper");
-            _behaviorEndTime = ResetBehaviorTime(Convert.ToSingle(BehaviorList.GetBehaviorStat("Stomper", "attackTime")));
-        }
-    }
-
-    public override void HandleStomper()
-    {
-        base.HandleStomper();
-
-        //DetectPlayers("Aggro");
-
-        if (Room.Time >= _behaviorEndTime)
-        {
-AiBehavior = ChangeBehavior("Patrol");
-            Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf("Patrol"), string.Empty, Position.x, Position.y, Generic.Patrol_ForceDirectionX, false));
+            AiBehavior = ChangeBehavior("Patrol");
         }
     }
 
     public override void DetectPlayers(string behaviorToRun)
     {
         foreach (var player in Room.Players)
-        {
             if (PlayerInRange(player.Value.TempData.Position, EnemyGlobalProps.Global_DetectionLimitedByPatrolLine))
             {
-                Room.SendSyncEvent(Utils.AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(behaviorToRun), string.Empty, player.Value.TempData.Position.X,
+                Room.SendSyncEvent(AISyncEventHelper.AIDo(Id, Room.Time, Position, 1.0f, BehaviorList.IndexOf(behaviorToRun), string.Empty, player.Value.TempData.Position.X,
                     Position.y, Generic.Patrol_ForceDirectionX, false));
 
                 // For some reason, the SyncEvent doesn't initialize these properly, so I just do them here
@@ -105,6 +98,5 @@ AiBehavior = ChangeBehavior("Patrol");
 
                 _behaviorEndTime = ResetBehaviorTime(MinBehaviorTime);
             }
-        }
     }
 }
