+++ Server.Reawakened/Rooms/Room.cs
@@ -7,6 +7,7 @@ using Server.Reawakened.Configs;
 using Server.Reawakened.Entities.Components;
 using Server.Reawakened.Entities.Entity;
 using Server.Reawakened.Entities.Entity.Enemies;
+using Server.Reawakened.Entities.Entity.Enemies.BehaviorEnemies;
 using Server.Reawakened.Entities.Interfaces;
 using Server.Reawakened.Network.Extensions;
 using Server.Reawakened.Players;
@@ -42,7 +43,7 @@ public class Room : Timer
 
     public Dictionary<string, PlaneModel> Planes;
     public Dictionary<string, List<string>> UnknownEntities;
-    public Dictionary<string, BehaviorEnemy> Enemies;
+    public Dictionary<string, Enemy> Enemies;
     public Dictionary<string, List<List<BaseComponent>>> DuplicateEntities;
 
     private readonly Dictionary<string, List<BaseComponent>> _entities;
@@ -151,6 +152,12 @@ public class Room : Timer
                         break;
                     case string vespid when vespid.Contains(config.EnemyNameSearch[11]):
                         Enemies.Add(component.Id, new EnemyVespid(this, component.Id, component.PrefabName, (EnemyControllerComp)component, services));
+                        break;
+                    case string spiderling when spiderling.Contains(config.EnemyNameSearch[12]):
+                        Enemies.Add(component.Id, new EnemySpiderling(this, component.Id, component.PrefabName, (EnemyControllerComp)component, services));
+                        break;
+                    case string rachnok when rachnok.Contains(config.EnemyNameSearch[13]):
+                        Enemies.Add(component.Id, new EnemyRachnok(this, component.Id, component.PrefabName, (EnemyControllerComp)component, services));
                         break;
                 }
             }
