+++ Web.AssetBundles/Assets/LocalAssets/InternalDefaultEnemies.xml
@@ -267,6 +267,25 @@
     </Enemy>
     
     <!--Rachnoids-->
+    <Enemy name="PF_Critter_Spider_Newb">
+        <GlobalProperties>
+            <OffensiveBehavior value="Aggro"/>
+            <MinBehaviorTime value="0.5"/>
+            <DetectionLimitedByPatrolLine value="true"/>
+            <FrontDetectionRangeX value="6"/>
+            <FrontDetectionRangeUpY value="1"/>
+            <FrontDetectionRangeDownY value="0.25"/>
+            <BackDetectionRangeX value="3"/>
+            <BackDetectionRangeUpY value="1"/>
+            <BackDetectionRangeDownY value="0"/>
+            <ViewOffsetY value="0"/>
+            <ProjectilePrefabName value=""/>
+        </GlobalProperties>
+        <Patrol speed="4.0" smoothMove="false" endPathWaitTime="2"/>
+        <Aggro speed="5.0" moveBeyondTargetDistance="0" stayOnPatrolPath="true" attackBeyondPatrolLine="0" detectionRangeUpY="0" detectionRangeDownY="0"/>
+        <LookAround lookTime="3" startDirection="0" forceDirection="0" initialProgressRatio="0" snapOnGround="false"/>
+        <Hitbox width="1.3" height="1" offsetX="0.5" offsetY="0"/>
+    </Enemy>
     <Enemy name="PF_Critter_Spider_Boss">
         <GlobalProperties>
             <OffensiveBehavior value="Aggro"/>
@@ -1911,5 +1930,17 @@
         <Stinger speedForward="12" speedBackward="4" inDurationForward="0.43" attackDuration="2.67" damageAttackTimeOffset="1" inDurationBackward="1.29" damageDistance="2">
             <Resource type="EABIL" name="PF_EABIL_Spite_Vespid_StingerAttack"/>
         </Stinger>
+    </Enemy>
+
+    <!--AI STATE ENEMIES BELOW THIS LINE-->
+
+    <!--Spiderling-->
+    <Enemy name="PF_Spite_Spiderling_Boss01">
+        <Hitbox width="1.3" height="1" offsetX="0" offsetY="0.5"/>
+    </Enemy>
+
+    <!--Rachnok-->
+    <Enemy name="Spite_TeaserSpider_Boss">
+        <Hitbox width="15" height="9" offsetX="0" offsetY="5.2"/>
     </Enemy>
 </InternalDefaultEnemies>
